+++ Packages/com.yontalane.core/Runtime/LeashTransform.cs
@@ -12,15 +12,25 @@ namespace Yontalane
             Manual = 1
         }
 
+        public enum Space
+        {
+            Local = 0,
+            World = 1
+        }
+
         [System.Serializable]
         public class Config
         {
+#if UNITY_EDITOR
+#pragma warning disable 0414 // Disable "private field assigned but not used" warning.
             [SerializeField]
             private string m_name = "";
+#endif
 
             public bool shouldLeash = default;
             public OffsetType offsetType = OffsetType.Auto;
             public Vector3 offset = default;
+            public Space space = Space.World;
 
             [Min(0f)]
             public float slack = 0f;
@@ -78,6 +88,9 @@ namespace Yontalane
         private Config m_rotationConfig = new Config(false);
 
         [SerializeField]
+        private Config m_scaleConfig = new Config(false);
+
+        [SerializeField]
         private UpdateType m_updateType = UpdateType.Update;
 
         private void Start()
@@ -95,6 +108,11 @@ namespace Yontalane
             if (m_rotationConfig.shouldLeash && m_rotationConfig.offsetType == OffsetType.Auto)
             {
                 m_rotationConfig.offset = transform.eulerAngles - Target.eulerAngles;
+            }
+
+            if (m_scaleConfig.shouldLeash && m_scaleConfig.offsetType == OffsetType.Auto)
+            {
+                m_scaleConfig.offset = transform.localScale - Target.localScale;
             }
         }
 
@@ -114,13 +132,60 @@ namespace Yontalane
 
             if (m_hasRigidbody)
             {
-                if (m_positionConfig.shouldLeash) Rigidbody.MovePosition(m_positionConfig.GetDestination(transform.position, Target.position));
-                if (m_rotationConfig.shouldLeash) Rigidbody.MoveRotation(Quaternion.Euler(m_rotationConfig.GetDestination(transform.eulerAngles, Target.eulerAngles)));
+                if (m_positionConfig.shouldLeash)
+                {
+                    if (m_positionConfig.space == Space.Local)
+                    {
+                        Vector3 v = m_positionConfig.GetDestination(transform.localPosition, Target.localPosition);
+                        Rigidbody.MovePosition(transform.TransformPoint(v));
+                    }
+                    else
+                    {
+                        Rigidbody.MovePosition(m_positionConfig.GetDestination(transform.position, Target.position));
+                    }
+                }
+                if (m_rotationConfig.shouldLeash)
+                {
+                    if (m_rotationConfig.space == Space.Local)
+                    {
+                        Vector3 v = m_rotationConfig.GetDestination(transform.eulerAngles, Target.eulerAngles);
+                        Rigidbody.MoveRotation(Quaternion.Euler(transform.TransformDirection(v)));
+                    }
+                    else
+                    {
+                        Rigidbody.MoveRotation(Quaternion.Euler(m_rotationConfig.GetDestination(transform.eulerAngles, Target.eulerAngles)));
+                    }
+                }
             }
             else
             {
-                if (m_positionConfig.shouldLeash) transform.position = m_positionConfig.GetDestination(transform.position, Target.position);
-                if (m_rotationConfig.shouldLeash) transform.eulerAngles = m_rotationConfig.GetDestination(transform.eulerAngles, Target.eulerAngles);
+                if (m_positionConfig.shouldLeash)
+                {
+                    if (m_positionConfig.space == Space.Local)
+                    {
+                        transform.localPosition = m_positionConfig.GetDestination(transform.localPosition, Target.localPosition);
+                    }
+                    else
+                    {
+                        transform.position = m_positionConfig.GetDestination(transform.position, Target.position);
+                    }
+                }
+                if (m_rotationConfig.shouldLeash)
+                {
+                    if (m_rotationConfig.space == Space.Local)
+                    {
+                        transform.localEulerAngles = m_rotationConfig.GetDestination(transform.localEulerAngles, Target.localEulerAngles);
+                    }
+                    else
+                    {
+                        transform.eulerAngles = m_rotationConfig.GetDestination(transform.eulerAngles, Target.eulerAngles);
+                    }
+                }
+            }
+
+            if (m_scaleConfig.shouldLeash)
+            {
+                transform.localScale = m_scaleConfig.GetDestination(transform.localScale, Target.localScale);
             }
         }
     }