+++ Packages/com.yontalane.core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.0.5 - 2022.05.17
+
+**Added**
+
+* Leash scale.
+* Leash local/world space option.
+
+**Changed**
+
+**Fixed**
+
 ## 1.0.4 - 2022.05.17
 
 **Added**